pyproject.toml
[project]
name = "avtodoc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.38.0",
    "pandoc>=2.4",
    "pypandoc>=1.15",
]
